requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.5.0
datasets>=2.14.0
trl>=0.7.0
sentence-transformers>=2.2.0
nltk>=3.8
pyyaml>=6.0
accelerate>=0.24.0
wandb>=0.15.0  # Optional, comment if no login
scikit-learn>=1.3.0
requests>=2.31.0
huggingface-hub>=0.17.0